Solved a bug in ImgFileLoader for resolutions higher than 2592x1944
Former-commit-id: 2c9ed571dea48ceacc834677b376eeb6cd3a1417

+++ AutoportLinux/src/ImgLoader/ImgFileLoader.cpp
@@ -75,13 +75,13 @@ void ImgFileLoader::getCropVector(Point2f &t) {
 
 bool ImgFileLoader::setResolutionWidth (int frameWidth)  {
 	res.width = frameWidth;
-	resampleMat.at<float>(0,0) = 2592/frameWidth;
+	resampleMat.at<float>(0,0) = 2592/(float)frameWidth;
 	return true;
 }
 
 bool ImgFileLoader::setResolutionHeight(int frameHeight) {
 	res.height = frameHeight;
-	resampleMat.at<float>(1,1) = 1944/frameHeight;
+	resampleMat.at<float>(1,1) = 1944/(float)frameHeight;
 	return true;
 }
